Be a little less agressive with vendor prefixes
I am not quite sure which browsers you intend to support but I believe it's safe to get rid of `-khtml-` prefixes as well as `-ms-` and `-moz-` prefixes for transitions (especially since the `-ms-` one has never been used). We could even ditch `-o-` I believe, but I left it just in case.

Feel free to decline if that does not suit your browser needs.